requirements.txt
beautifulsoup4==4.10.0
Flask==2.0.2
flask_restx==0.5.1
requests==2.26.0
PyYAML==6.0